Implemented Event System and added containers for dynamic arrays.
The event system is very lite and rudimentary, however it does open the door for input and other sub-systems down the line. The current implementation is single threaded and do not allow for multi-threading however, that is due to change.

Furthermore, the new containers system allows for dynamic arrays which will be so useful & used internally all over the engine! :)